secured endpoints

+++ app/builds/routes.py
@@ -4,6 +4,7 @@ from app.builds.models import Build as BuildModel
 from app.builds.models import BuildItem
 from app.builds.schemas import Build as BuildSchema
 from app.builds.schemas import BuildResponse
+from app.core.security import get_current_user
 from app.db.database import get_db
 from app.items.models import Item
 from app.users.models import User
@@ -19,17 +20,26 @@ def get_builds(db: Session = Depends(get_db)):
     return builds
 
 
-@router.post("/", response_model=BuildResponse, status_code=status.HTTP_201_CREATED)
-def create_build(build: BuildSchema, db: Session = Depends(get_db)):
-
-    if not db.query(User).get(build.user_id):
+@router.get("/{name}/builds", response_model=List[BuildResponse])
+def get_user_builds(name: str, db: Session = Depends(get_db)):
+    user = db.query(User).filter_by(name=name).first()
+    if not user:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="No user found for given user_id",
+            detail="User not found",
         )
+    builds = db.query(BuildModel).filter_by(user=user).all()
+    return builds
 
+
+@router.post("/", response_model=BuildResponse, status_code=status.HTTP_201_CREATED)
+def create_build(
+    build: BuildSchema,
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
     new_build = BuildModel(
-        name=build.name, description=build.description, user_id=build.user_id
+        name=build.name, description=build.description, user_id=current_user.id
     )
     item_list = []
     for item_id in build.items:
@@ -45,12 +55,23 @@ def create_build(build: BuildSchema, db: Session = Depends(get_db)):
 
 
 @router.delete("/{build_id}", response_model=BuildResponse)
-def remove_build(build_id: int, db: Session = Depends(get_db)):
+def remove_build(
+    build_id: int,
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
     build = db.query(BuildModel).get(build_id)
+
     if not build:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Build not found"
         )
+
+    if not build.user == current_user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized"
+        )
+
     db.delete(build)
     db.commit()
     return build
